Create ResourceTab component
