+++ src/pages/MirrorPage.tsx
@@ -1,25 +1,13 @@
-import { stringify } from '@medplum/core';
 import { Resource, OperationOutcome, Bundle, Questionnaire } from '@medplum/fhirtypes';
 import {
-  Button,
-  DefaultResourceTimeline,
   Document,
-  EncounterTimeline,
   ErrorBoundary,
-  Form,
   Loading,
   MedplumLink,
-  PatientTimeline,
-  ResourceBlame,
-  ResourceForm,
-  ResourceHistoryTable,
-  ResourceTable,
-  ServiceRequestTimeline,
   Tab,
   TabList,
   TabPanel,
   TabSwitch,
-  TextArea,
   useMedplum,
 } from '@medplum/react';
 import React, { useState, useCallback, useEffect } from 'react';
@@ -27,6 +15,7 @@ import { useParams, useNavigate } from 'react-router-dom';
 import './MirrorPage.css';
 import { PatientHeader } from './PatientHeader';
 import { ResourceHeader } from './ResourceHeader';
+import { ResourceTab } from './ResourceTab';
 import { getPatient } from './utils';
 
 export function getTabs(resourceType: string): string[] {
@@ -175,78 +164,4 @@ export function MirrorPage(): JSX.Element {
       )}
     </>
   );
-}
-
-interface ResourceTabProps {
-  name: string;
-  resource: Resource;
-  resourceHistory: Bundle;
-  questionnaires: Bundle<Questionnaire>;
-  onSubmit: (resource: Resource) => void;
-  outcome?: OperationOutcome;
-}
-
-function ResourceTab(props: ResourceTabProps): JSX.Element | null {
-  const navigate = useNavigate();
-  const medplum = useMedplum();
-  const { resourceType, id } = props.resource;
-  switch (props.name) {
-    case 'details':
-      return <ResourceTable value={props.resource} />;
-    case 'history':
-      return <ResourceHistoryTable history={props.resourceHistory} />;
-    case 'blame':
-      return <ResourceBlame history={props.resourceHistory} />;
-    case 'timeline':
-      if (props.resource.resourceType === 'Patient') {
-        return <PatientTimeline patient={props.resource} />;
-      }
-      if (props.resource.resourceType === 'Encounter') {
-        return <EncounterTimeline encounter={props.resource} />;
-      }
-      if (props.resource.resourceType === 'ServiceRequest') {
-        return <ServiceRequestTimeline serviceRequest={props.resource} />;
-      }
-      return <DefaultResourceTimeline resource={props.resource} />;
-    case 'edit':
-      return (
-        <ResourceForm
-          defaultValue={props.resource}
-          onSubmit={props.onSubmit}
-          onDelete={() => navigate(`/${resourceType}/${id}/delete`)}
-        />
-      );
-    case 'delete':
-      return (
-        <>
-          <p>Are you sure you want to delete this {resourceType}?</p>
-          <Button
-            danger={true}
-            onClick={() => {
-              medplum.deleteResource(resourceType, id as string).then(() => navigate(`/${resourceType}`));
-            }}
-          >
-            Delete
-          </Button>
-        </>
-      );
-    case 'json':
-      return (
-        <Form
-          onSubmit={(formData: Record<string, string>) => {
-            props.onSubmit(JSON.parse(formData.resource));
-          }}
-        >
-          <TextArea
-            testid="resource-json"
-            name="resource"
-            monospace={true}
-            style={{ height: 400 }}
-            defaultValue={stringify(props.resource, true)}
-          />
-          <Button type="submit">OK</Button>
-        </Form>
-      );
-  }
-  return null;
 }
